requirements.txt
streamlit
numpy
Pillow
reportlab
pandas
torch
torchvision